refactor(app-shell): extract authenticated layout and nav chrome
